Circles for the win
Road and flooding is done.

Changes to grid.py:
Made Grid.get_active_tiles function for counting all active tiles in the road.

Changes to main.py
Added the flood() function that grants values to all active tiles based on distance from the finish line and made adjustments to the road() function for visualizing the finish line. Also made adjustments to the main() loop for better performance and cooperation with other functions.